+++ Scripts/CustomActions/Remove-NavEnvironment.ps1
@@ -5,8 +5,11 @@ if ($BranchSettings.instanceName -ne "") {
         Remove-NAVEnvironment -BranchSettings $BranchSettings
         UnLoad-InstanceAdminTools
     } else {
-        docker.exe kill $($BranchSettings.dockerContainerName)
-        docker.exe rm $($BranchSettings.dockerContainerName)
+        $dockerContainer = Get-DockerContainers | Where-Object -Property Id -ieq $BranchSettings.dockerContainerName
+        if ($dockerContainer) {
+            docker.exe kill $($BranchSettings.dockerContainerName)
+            docker.exe rm $($BranchSettings.dockerContainerName)
+        }
         $BranchSettings = Clear-BranchSettings -BranchId $BranchSettings.branchId 
     }
 }
